requirements.txt
slack-sdk==3.31.0
slack-bolt==1.19.1
python-dotenv==1.0.1
aiohttp==3.9.5
pydantic==2.7.4
redis==5.0.6
celery==5.4.0
requests==2.31.0
llama-cpp-python==0.2.90
huggingface-hub==0.24.5